Fix scripts idempotency

+++ schema/reportdb/upgrade/uyuni-reportdb-schema-4.3.1-to-uyuni-reportdb-schema-4.3.2/001-change-event-data-type.sql
@@ -1,31 +1,39 @@
-DROP VIEW IF EXISTS HistoryReport;
+DO $$
+  BEGIN
+    IF (SELECT data_type FROM information_schema.columns WHERE table_name = 'historyreport' AND column_name = 'event_data') != 'text' THEN
+      DROP VIEW IF EXISTS HistoryReport;
 
-ALTER TABLE SystemAction ALTER COLUMN event_data SET DATA TYPE TEXT;
+      ALTER TABLE SystemAction ALTER COLUMN event_data SET DATA TYPE TEXT;
 
-ALTER TABLE SystemHistory ALTER COLUMN event_data SET DATA TYPE TEXT;
+      ALTER TABLE SystemHistory ALTER COLUMN event_data SET DATA TYPE TEXT;
 
-CREATE OR REPLACE VIEW HistoryReport AS
-      SELECT mgm_id
-                , system_id
-                , action_id AS event_id
-                , hostname
-                , event
-                , completion_time AS event_time
-                , status
-                , event_data
-                , synced_date
-        FROM SystemAction
+      CREATE OR REPLACE VIEW HistoryReport AS
+            SELECT mgm_id
+                      , system_id
+                      , action_id AS event_id
+                      , hostname
+                      , event
+                      , completion_time AS event_time
+                      , status
+                      , event_data
+                      , synced_date
+              FROM SystemAction
 
-    UNION ALL
+          UNION ALL
 
-      SELECT mgm_id
-                , system_id
-                , history_id AS event_id
-                , hostname
-                , event
-                , event_time
-                , 'Done' AS status
-                , event_data
-                , synced_date
-        FROM SystemHistory
-;
+            SELECT mgm_id
+                      , system_id
+                      , history_id AS event_id
+                      , hostname
+                      , event
+                      , event_time
+                      , 'Done' AS status
+                      , event_data
+                      , synced_date
+              FROM SystemHistory
+      ;
+    ELSE
+      RAISE NOTICE 'systemaction column event_data is already of type text.';
+    END IF;
+  END;
+$$;
